requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.13.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.2
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tldextract==5.1.1
tzdata==2023.3
urllib3==2.1.0
zope.event==5.0
zope.interface==6.1
